users/tests/test_biometric_utils.py
"""
Unit tests for biometric utility functions.

Tests cover:
- Face detection and extraction
- Face encoding generation
- Face comparison and matching
- ID card OCR text extraction
- Image preprocessing
- Face quality validation
"""

from django.test import TestCase
from django.core.files.uploadedfile import SimpleUploadedFile
from django.conf import settings
from PIL import Image
import numpy as np
import io
import os

from users.biometric_utils import (
    load_image_from_file,
    preprocess_id_image,
    extract_text_from_id,
    detect_and_extract_face,
    generate_face_encoding,
    compare_faces,
    validate_face_quality,
    encoding_to_json,
    json_to_encoding,
    get_face_match_threshold,
    BiometricError,
    FaceDetectionError,
    FaceQualityError,
    OCRError
)


class BiometricUtilsTestCase(TestCase):
    """Test cases for biometric utility functions"""
    
    def setUp(self):
        """Set up test fixtures"""
        # Create a simple test image (100x100 white image)
        self.test_image = Image.new('RGB', (100, 100), color='white')
        self.test_image_bytes = io.BytesIO()
        self.test_image.save(self.test_image_bytes, format='JPEG')
        self.test_image_bytes.seek(0)
        
        # Create uploaded file
        self.uploaded_file = SimpleUploadedFile(
            "test_image.jpg",
            self.test_image_bytes.read(),
            content_type="image/jpeg"
        )
        self.test_image_bytes.seek(0)
    
    def test_load_image_from_file(self):
        """Test loading image from file"""
        try:
            image = load_image_from_file(self.uploaded_file)
            self.assertIsInstance(image, np.ndarray)
            self.assertEqual(len(image.shape), 3)  # Should be RGB (3 channels)
        except BiometricError:
            # Expected if face_recognition has issues with test image
            pass
    
    def test_preprocess_id_image(self):
        """Test ID card image preprocessing"""
        # Create a simple RGB image
        test_array = np.ones((100, 100, 3), dtype=np.uint8) * 255
        
        processed = preprocess_id_image(test_array)
        
        # Check that output is grayscale
        self.assertEqual(len(processed.shape), 2)
        self.assertIsInstance(processed, np.ndarray)
    
    def test_encoding_to_json_and_back(self):
        """Test converting face encoding to JSON and back"""
        # Create a mock 128-dimensional encoding
        mock_encoding = np.random.rand(128)
        
        # Convert to JSON
        json_data = encoding_to_json(mock_encoding)
        self.assertIsInstance(json_data, list)
        self.assertEqual(len(json_data), 128)
        
        # Convert back to numpy array
        recovered_encoding = json_to_encoding(json_data)
        self.assertIsInstance(recovered_encoding, np.ndarray)
        self.assertEqual(len(recovered_encoding), 128)
        
        # Check values are preserved
        np.testing.assert_array_almost_equal(mock_encoding, recovered_encoding)
    
    def test_get_face_match_threshold(self):
        """Test getting face match threshold from settings"""
        threshold = get_face_match_threshold()
        self.assertIsInstance(threshold, float)
        self.assertGreater(threshold, 0)
        self.assertLessEqual(threshold, 1.0)
    
    def test_compare_faces_with_mock_encodings(self):
        """Test face comparison with mock encodings"""
        # Create two identical encodings
        encoding1 = np.random.rand(128)
        encoding2 = encoding1.copy()
        
        is_match, distance = compare_faces(encoding1, encoding2, tolerance=0.6)
        
        # Identical encodings should have distance of 0
        self.assertEqual(distance, 0.0)
        self.assertTrue(is_match)
        
        # Create two different encodings
        encoding3 = np.random.rand(128)
        encoding4 = np.random.rand(128)
        
        is_match2, distance2 = compare_faces(encoding3, encoding4, tolerance=0.6)
        
        # Different encodings should have non-zero distance
        self.assertGreater(distance2, 0.0)
    
    def test_compare_faces_with_list_input(self):
        """Test face comparison accepts list input"""
        # Create encodings as lists
        encoding1 = [0.1] * 128
        encoding2 = [0.1] * 128
        
        is_match, distance = compare_faces(encoding1, encoding2, tolerance=0.6)
        
        # Should handle list input
        self.assertIsInstance(distance, float)
        self.assertIsInstance(is_match, bool)
    
    def test_face_detection_error_on_invalid_image(self):
        """Test that FaceDetectionError is raised for invalid images"""
        # This test would require actual face detection
        # For now, we just verify the exception exists
        self.assertTrue(issubclass(FaceDetectionError, BiometricError))
    
    def test_biometric_error_hierarchy(self):
        """Test exception hierarchy"""
        self.assertTrue(issubclass(FaceDetectionError, BiometricError))
        self.assertTrue(issubclass(FaceQualityError, BiometricError))
        self.assertTrue(issubclass(OCRError, BiometricError))


class BiometricIntegrationTestCase(TestCase):
    """Integration tests for biometric features"""
    
    def test_biometric_settings_exist(self):
        """Test that biometric settings are configured"""
        self.assertTrue(hasattr(settings, 'BIOMETRIC_SETTINGS'))
        
        biometric_settings = settings.BIOMETRIC_SETTINGS
        self.assertIn('FACE_MATCH_THRESHOLD', biometric_settings)
        self.assertIn('MIN_FACE_SIZE', biometric_settings)
        self.assertIn('MAX_IMAGE_SIZE', biometric_settings)
        self.assertIn('FACE_RECOGNITION_ENABLED', biometric_settings)
    
    def test_face_match_threshold_value(self):
        """Test face match threshold is reasonable"""
        threshold = settings.BIOMETRIC_SETTINGS['FACE_MATCH_THRESHOLD']
        self.assertGreaterEqual(threshold, 0.0)
        self.assertLessEqual(threshold, 1.0)
    
    def test_min_face_size_configured(self):
        """Test minimum face size is configured"""
        min_size = settings.BIOMETRIC_SETTINGS['MIN_FACE_SIZE']
        self.assertIsInstance(min_size, tuple)
        self.assertEqual(len(min_size), 2)
        self.assertGreater(min_size[0], 0)
        self.assertGreater(min_size[1], 0)


# Note: Full integration tests with actual face images would require:
# 1. Sample face images in a test fixtures directory
# 2. Sample ID card images
# 3. face_recognition library properly installed
# 4. Tesseract OCR installed on the system
#
# These tests focus on unit testing the utility functions with mock data.
# For production, you should add integration tests with real images.
