requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.11.0
attrs==25.4.0
bitarray==3.8.0
certifi==2026.1.4
charset-normalizer==3.4.4
ckzg==2.1.5
cytoolz==1.1.0
Django==4.2.9
django-cors-headers==4.3.1
django-ratelimit==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
frozenlist==1.8.0
hexbytes==1.3.1
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
multidict==6.7.0
node==1.2.3
odict==1.9.0
packaging==25.0
parsimonious==0.10.0
pillow==12.1.0
plumber==1.7
propcache==0.4.1
psycopg2-binary==2.9.11
py-solc-x==2.0.5
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
PyPDF2==3.0.1
python-decouple==3.8
pytz==2025.2
pyunormalize==17.0.0
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
reportlab==4.0.7
requests==2.32.5
rlp==4.1.0
rpds-py==0.30.0
setuptools==80.9.0
sqlparse==0.5.5
toolz==1.1.0
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
web3==7.14.0
websockets==15.0.1
yarl==1.22.0
zope.component==7.0
zope.deferredimport==6.0
zope.deprecation==6.0
zope.event==6.1
zope.hookable==8.2
zope.interface==8.2
zope.lifecycleevent==6.0
zope.proxy==7.1
python-dotenv
pytesseract==0.3.13
opencv-python==4.13.0.90
dlib-bin==20.0.0  # Pre-built dlib binary (no C++ compiler needed)
face-recognition==1.2.3  # Install with --no-deps to avoid pulling dlib
face-recognition-models==0.3.0